functions to save/edit, add, and delete extracted into new file
